aucarvideo/requirements.txt
Django==2.1
django-tenant-schemas==1.9.0
pkg-resources==0.0.0
psycopg2==2.7.5
pytz==2018.5
